Adding Show alert confirmation

+++ src/main/java/views/EmployeeListViewController.java
@@ -3,6 +3,7 @@ package main.java.views;
 import java.io.IOException;
 import java.net.URL;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -13,6 +14,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -116,10 +118,19 @@ public class EmployeeListViewController implements Initializable {
 
     try {
 
-      UserController.getInstance()
-          .deleteEmployeeById(
-              tableViewEmployee.getSelectionModel().getSelectedItem().getRegisterId());
-      updateTableView();
+      if (tableViewEmployee.getSelectionModel().getSelectedItem().getRegisterId() != 0) {
+
+        Optional<ButtonType> result =
+            Alerts.showConfirmation(
+                "Confirmation", "Are you sure you want to delete the selected field ?");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+
+          UserController.getInstance()
+              .deleteEmployeeById(
+                  tableViewEmployee.getSelectionModel().getSelectedItem().getRegisterId());
+          updateTableView();
+        }
+      }
     } catch (Exception e) {
       Alerts.showAlert("Error", null, "Select the field to be deleted", AlertType.ERROR);
     }
