+++ src/main/java/views/ProductListViewController.java
@@ -3,6 +3,7 @@ package main.java.views;
 import java.io.IOException;
 import java.net.URL;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -13,6 +14,7 @@ import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -111,9 +113,18 @@ public class ProductListViewController implements Initializable {
 
     try {
 
-      ProductController.getInstance()
-          .deleteProductById(tableViewProduct.getSelectionModel().getSelectedItem().getId());
-      updateTableView();
+      if (tableViewProduct.getSelectionModel().getSelectedItem().getId() != 0) {
+
+        Optional<ButtonType> result =
+            Alerts.showConfirmation(
+                "Confirmation", "Are you sure you want to delete the selected field ?");
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+
+          ProductController.getInstance()
+              .deleteProductById(tableViewProduct.getSelectionModel().getSelectedItem().getId());
+          updateTableView();
+        }
+      }
     } catch (Exception e) {
       Alerts.showAlert("Error", null, "Select the field to be deleted", AlertType.ERROR);
     }
